handle click outside of bounds
